fixup! refactor: change package name to standard Java naming
