+++ src/main/java/com/crimsoncrips/borninconfiguration/mixins/spawning/maggot/ZombieClownMaggot.java
@@ -1,4 +1,4 @@
-package com.crimsoncrips.borninconfiguration.mixins.spawning.Maggots;
+package com.crimsoncrips.borninconfiguration.mixins.spawning.maggot;
 
 
 import com.crimsoncrips.borninconfiguration.config.BIConfig;
@@ -26,7 +26,7 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(ZombieClownDeathTimeIsReachedProcedure.class)
 
-public class ZombieClown_Maggot {
+public class ZombieClownMaggot {
 
     @Inject(method = "execute", at = @At("HEAD"), cancellable = true, remap = false)
     private static void injected(LevelAccessor world, double x, double y, double z, CallbackInfo ci) {
